Adding second "current_is_async" method in "async.c"
In this commit i copied the duplicated the method "current_is_async" in "kernel/async.c", but the new method is "current_is_async_pars". So the method without "pars"
is directly checking if the current task is asynchronous. And the new method is wait for a parameter of the "worker" struct, and checks then if the given "worker" struct
is asynchronous.

author: <leotendosteinhoff@gmail.com>
commiter: <leotendosteinhoff@gmail.com>
branch: <master>
date: 28th of Nov, 12:28